# docs/swagger.yaml
basePath: /
definitions:
...
  request.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  request.RefreshReq:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  request.RegisterReq:
    properties:
      code:
        maxLength: 6
        minLength: 6
        type: string
      email:
        maxLength: 50
        type: string
      name:
        maxLength: 20
        minLength: 3
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - code
    - email
    - name
    - password
    type: object
  response.LoginResp:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
  response.Result:
    properties:
      code:
        description: 业务码：0 成功，非 0 失败
        type: integer
      data:
        description: 业务数据
      message:
        description: 提示信息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: 1473443474@qq.com
    name: API Support
  description: DisembodiedSpecter API 文档
  termsOfService: http://swagger.io/terms/
  title: DisembodiedSpecter API
  version: "1.0"
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 通过用户名密码进行登录，返回双令牌（access_token + refresh_token）
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回双令牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Result'
            - properties:
                data:
                  $ref: '#/definitions/response.LoginResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Result'
      summary: 用户登录
      tags:
      - 认证模块
  /api/v1/logout:
    post:
      consumes:
      - application/json
      description: 退出登录接口（需要 access token 和 refresh token）
      parameters:
      - description: 退出登录请求参数（refresh_token）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RefreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: 退出登录成功
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Result'
      summary: 用户退出登录接口
      tags:
      - 认证模块
  /api/v1/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh token 获取新的双令牌
      parameters:
      - description: 刷新令牌请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RefreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回新的双令牌
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Result'
      summary: 刷新令牌
      tags:
      - 认证模块
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: 注册新用户，返回双令牌（access_token + refresh_token）
      parameters:
      - description: 注册请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回双令牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Result'
            - properties:
                data:
                  $ref: '#/definitions/response.LoginResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Result'
      summary: 用户注册
      tags:
      - 认证模块
  /api/v1/send_verify_code:
    get:
      consumes:
      - application/json
      description: 发送验证码到指定邮箱
      parameters:
      - description: 验证目标（邮箱）
        in: query
        name: verifier
        required: true
        type: string
      - description: 验证类型（email）
        in: query
        name: verifier_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 发送验证码成功
          schema:
            $ref: '#/definitions/response.Result'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Result'
      summary: 发送验证码接口
      tags:
      - 认证模块
  /api/v2/data:
    get:
      consumes:
      - application/json
      description: 获取用户数据，包括邮箱、状态等
      produces:
      - application/json
      responses:
        "200":
          description: '成功 data: 用户数据对象'
          schema:
            $ref: '#/definitions/response.Result'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/response.Result'
      summary: 获取用户数据
      tags:
      - 用户
  /api/v2/ws-code:
    get:
      consumes:
      - application/json
      description: 获取 WebSocket 连接码，用于连接 WebSocket 服务器
      produces:
      - application/json
      responses:
        "200":
          description: '成功 data: {"wsCode": "123456"}'
          schema:
            $ref: '#/definitions/response.Result'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/response.Result'
      summary: 获取 WebSocket 连接码
      tags:
      - 用户
...
swagger: "2.0"